requirements.txt
alembic==1.13.2
apturl==0.5.2
asgiref==3.8.1
attrs==22.2.0
Authlib==1.3.1
azure-common==1.1.28
azure-core==1.30.2
azure-identity==1.17.1
azure-mgmt-core==1.4.0
azure-mgmt-rdbms==10.1.0
azure-mgmt-resource==23.1.1
azure-mgmt-subscription==3.1.1
babel==2.16.0
backports.tarfile==1.2.0
bcrypt==4.2.0
bidict==0.23.1
blinker==1.8.2
boto3==1.34.162
botocore==1.34.162
Brlapi==0.8.3
Brotli==1.1.0
CacheControl==0.12.10
cachetools==5.5.0
cachy==0.3.0
certifi==2020.6.20
cffi==1.17.0
chardet==4.0.0
cleo==0.8.1
click==8.1.7
clikit==0.6.2
colorama==0.4.6
command-not-found==0.3
coverage==7.0.5
crashtest==0.3.1
cryptography==43.0.0
cupshelpers==1.0
dbus-python==1.2.18
defer==1.0.6
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1
djangorestframework==3.15.2
dnspython==2.6.1
duplicity==0.8.21
email_validator==2.2.0
eventlet==0.36.1
exceptiongroup==1.2.1
fasteners==0.14.1
filelock==3.6.0
Flask==3.0.3
flask-babel==4.0.0
Flask-Compress==1.15
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Migrate==4.0.7
Flask-Paranoid==0.3.0
Flask-Principal==0.4.0
Flask-Security-Too==5.5.2
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
future==0.18.2
google-api-core==2.19.1
google-api-python-client==2.142.0
google-auth==2.34.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.63.2
greenlet==1.1.2
h11==0.14.0
html5lib==1.1
httplib2==0.20.2
idna==3.3
importlib_metadata==8.4.0
importlib_resources==6.4.4
iniconfig==2.0.0
isodate==0.6.1
itsdangerous==2.2.0
jaraco.classes==3.4.0
jaraco.context==6.0.1
jaraco.functools==4.0.2
jeepney==0.7.1
Jinja2==3.1.4
jmespath==1.0.1
jsonformatter==0.3.2
keyring==25.3.0
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
ldap3==2.9.1
libgravatar==1.0.4
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
monotonic==1.6
more-itertools==8.10.0
msal==1.30.0
msal-extensions==1.2.0
msgpack==1.0.3
msrest==0.7.1
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
packaging==23.0
PAM==0.4.2
paramiko==2.9.3
passlib==1.7.4
pastel==0.2.1
pexpect==4.8.0
pgadmin4==8.11
Pillow==9.0.1
pkginfo==1.8.2
platformdirs==2.5.1
pluggy==1.0.0
poetry==1.1.12
poetry-core==1.0.7
portalocker==2.10.1
proto-plus==1.24.0
protobuf==5.27.3
psutil==6.0.0
psycopg==3.2.1
psycopg-binary==3.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
Pygments==2.18.0
PyGObject==3.42.1
PyICU==2.8.1
PyJWT==2.3.0
pylev==1.2.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyotp==2.9.0
pyparsing==2.4.7
pypng==0.20220715.0
PyQt5==5.15.6
PyQt5-sip==12.9.1
pyRFC3339==1.1
pytest==7.2.1
pytest-cov==4.0.0
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-engineio==4.9.1
python-socketio==5.11.3
pytz==2024.1
pyxdg==0.27
PyYAML==5.4.1
qrcode==7.4.2
reportlab==3.6.8
requests==2.25.1
requests-oauthlib==2.0.0
requests-toolbelt==0.9.1
rich==13.7.1
rsa==4.9
s3transfer==0.10.2
SecretStorage==3.3.1
shellingham==1.4.0
simple-websocket==1.0.0
six==1.16.0
speaklater3==1.4
SQLAlchemy==2.0.32
sqlparse==0.4.2
sshtunnel==0.4.0
systemd-python==234
tomli==2.0.1
tomlkit==0.9.2
typer==0.12.4
typing_extensions==4.12.2
ua-parser==0.18.0
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
usb-creator==0.3.7
user-agents==2.2.0
virtualenv==20.13.0+ds
wadllib==1.3.6
webencodings==0.5.1
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
xdg==5
xkit==0.0.0
zipp==1.0.0
zstandard==0.23.0
